refactor(jobs): integrate agentService for job starting logic and enhance error handling
