+++ web-app/src/lib/services/user.service.ts
@@ -22,6 +22,19 @@ export const userService = {
   },
 
   /**
+   * Retrieves the active organization ID for the currently authenticated user.
+   *
+   * - Returns the organization ID if the user has an active organization in their session.
+   * - Returns null or undefined if there is no active organization or no session.
+   *
+   * @returns {Promise<string | null | undefined>} The active organization ID, or null/undefined if not set.
+   */
+  async getActiveOrganizationId(): Promise<string | null | undefined> {
+    const session = await getSession();
+    return session?.session.activeOrganizationId;
+  },
+
+  /**
    * Retrieves jobs for the currently authenticated user filtered by agent ID.
    * If the user has an active organization, returns jobs associated with that organization.
    * Otherwise, returns personal jobs for the user and agent.
